refactor(core): split blocks.rs into a blocks/ module directory
blocks.rs had grown to 2446 lines mixing several distinct
responsibilities. Split along the file's existing internal seams into
blocks/{mod,table_cell,markdown_text,tag_conversion,tests}.rs: no
behavior change, pure code motion with mod/use/visibility plumbing.
